Raise exception on log post failure
Catching all exceptions when posting logs caused the buffer to be
reset even when the post to the Monasca Log API failed. Therefore
if the post failed, the logs in the buffer would never be resent
to the Monasca Log API. By ensuring than an exception is raised,
the buffer will not be wiped when a post fails, and instead the
retry mechanism will be activated.

closes #2